Task/FP-424: Remove unused Manage Account code (#200)
* Removed unused templates and views

Moved form field data to lib

* Fixed AppRouter test

Moved form_fields back to apps.accounts package

* Remove login view

* Remove captcha comments

Removed logging

* Message test warning fix

* Remove call to messages.error

* Hide empty fields
