+++ client/src/components/Workbench/index.js
@@ -1,6 +1,6 @@
 import { hot } from 'react-hot-loader/root';
 import React from 'react';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Redirect } from 'react-router-dom';
 import Workbench from './Workbench';
 import ManageAccount from '../ManageAccount';
 import * as ROUTES from '../../constants/routes';
@@ -12,6 +12,7 @@ function AppRouter() {
       <Route path="/accounts/profile" component={ManageAccount} />
       <Route path={ROUTES.WORKBENCH} component={Workbench} />
       <Route path="/tickets/new" component={TicketUnauthenticatedCreate} />
+      <Redirect from="/accounts" to="/accounts/profile" />
     </Router>
   );
 }
