--- server/portal/apps/accounts/templates/portal/apps/accounts/department_dropdown_options.html
@@ -1,4 +1,0 @@
-<option value="" disabled selected>Choose one</option>
-{% for department in departments %}
-<option value="{{ department.id }}">{{ department.name }}</option>
-{% endfor %}
