--- server/portal/apps/accounts/templates/portal/apps/accounts/index.html
@@ -1,27 +1,0 @@
-{% extends "base.html" %}
-{% load sekizai_tags staticfiles %}
-{% block title %} Accounts {% endblock %}
-{% block content %}
-<div class="container">
-  <h1 class="headline headline-research">{{title}}</h1>
-  <div class="row">
-    <div class="col-sm-3">
-      <ul class="nav flex-column nav-pills" role="tablist" aria-orientation="vertical">
-        {% url 'portal_accounts:manage_profile' as item_url %}
-        <li class="nav-item">
-          <a class="nav-link {% if item_url in request.path %}active {% endif %}" href="{{item_url}}">Account Profile</a>
-        </li>
-        {% url 'portal_accounts:manage_authentication' as item_url %}
-        <li class="nav-item">
-          <a class="nav-link {% if item_url in request.path %}active {% endif %}" href="{{item_url}}">Authentication</a>
-        </li>
-      </ul>
-    </div>
-
-    <div class="col-sm-9">
-      {% block panel_content %} PLACEHOLDER {% endblock %}
-    </div>
-
-  </div>
-</div>
-{% endblock %}
