--- server/portal/apps/accounts/templates/portal/apps/accounts/manage_applications.html
@@ -1,26 +1,0 @@
-{% extends "portal/apps/accounts/index.html" %}
-
-{% block panel_content %}
-<div class="card">
-    <div class="card-body">
-        <h2>3rd Party Applications</h2>
-        <p class="lead">
-            Select the application settings to manage.
-        </p>
-        {% for app in integrations %}
-        <hr>
-        <div>
-            <h3>{{app.label}}</h3>
-            <p>{{app.description}}</p>
-            <p>
-                {% if app.href %}
-                    <a href="{{app.href}}" class="btn btn-default">Manage App Settings</a>
-                {% else %}
-                    <button disabled class="btn btn-default">Coming Soon</button>
-                {% endif %}
-            </p>
-        </div>
-        {% endfor %}
-    </div>
-</div>
-{% endblock %}
