--- server/portal/apps/accounts/templates/portal/apps/accounts/manage_auth.html
@@ -1,26 +1,0 @@
-{% extends "portal/apps/accounts/index.html" %}
-{% load bootstrap4 %}
-{% block panel_content %}
-<div class="card">
-    <div class="card-body">
-        <h2>Change Your Password</h2>
-        <hr>
-        <div class="alert alert-warning">
-            <h4>Changing your TACC Password</h4>
-            <strong>Please Note:</strong> This form requests a reset to your TACC Account
-            password. Changes to this password will affect your TACC Account
-            <em>as a whole</em>.
-        </div>
-
-        <form method="post" action="">
-            {% csrf_token %}
-
-            {% bootstrap_form form %}
-
-            {% buttons %}
-            <button type="submit" class="btn btn-primary">Change Password</button>
-            {% endbuttons %}
-        </form>
-    </div>
-</div>
-{% endblock %}
