--- server/portal/apps/accounts/templates/portal/apps/accounts/manage_licenses.html
@@ -1,33 +1,0 @@
-{% extends "portal/apps/accounts/index.html" %} {% block panel_content %}
-<div class="card">
-    <div class="card-body">
-        <h2>Software Licenses</h2>
-
-        <p class="lead">
-            This Portal makes a number of software programs available to users, some of which are commercial and require licenses to use.
-            Below you can review and find details on how to obtain and/or activate your license to use licensed software.
-        </p>
-
-        {% for license in licenses %}
-        <hr>
-        <div class="license {% if license.current_user_license %}license-active{% endif %}">
-            <h3>
-                {{license.license_type}} {% if not license.current_user_license %}
-                <span class="badge badge-success"><i class="fas fa-check-circle"></i> ACTIVE </span> {% else %}
-                <span class="badge badge-danger"> NOT ACTIVE </span> {% endif %}
-            </h3>
-            {% include license.details_html %} {% if not license.current_user_license %}
-            <p>
-                <a class="btn btn-lg btn-primary" href="{% url 'tickets:create' %}?category=OTHER&subject={{license.license_type}}+license+activation+request{% if license.license_type == 'LS-DYNA' %}&problem_description=Department+Name%3A%0DInstitution+Name%3A{% endif %}"
-                    target="_blank">
-                            Request Activation
-                        </a>
-            </p>
-            {% else %}
-            <p class="text-success">{{license.license_type}} is activated for your account.</p>
-            {% endif %}
-        </div>
-        {% endfor %}
-    </div>
-</div>
-{% endblock %}
