--- server/portal/apps/accounts/templates/portal/apps/accounts/manage_notifications.html
@@ -1,53 +1,0 @@
-{% extends "portal/apps/accounts/index.html" %}
-{% load bootstrap4 %}
-{% block panel_content %}
-
-<div class="card">
-    <div class="card-body">
-        <h2>Notifications</h2>
-        <hr>
-        <form method="post" action="">
-            {% csrf_token %}
-            <div class="row">
-                <div class="col-md-6">
-                    <h3>Email Notifications</h3>
-                    <p>The portal will notify you via email for the following:</p>
-                    {% bootstrap_form form %}
-                    <div class="form-group">
-                        <div class="checkbox">
-                            <label disabled>
-                                <input type="checkbox" checked disabled>
-                                Account updates such as password changes and profile updates
-                            </label>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-md-6">
-                    <h3>Site Notifications</h3>
-                    <p>The portal will notify you via in-site notifications for the following:</p>
-                    <div class="form-group">
-                        <div class="checkbox">
-                            <label>
-                                <input type="checkbox" checked disabled>
-                                Applications Workspace Job Status changes
-                            </label>
-                        </div>
-                    </div>
-                    <div class="form-group">
-                        <div class="checkbox">
-                            <label>
-                                <input type="checkbox" checked disabled>
-                                Files sharing notifications
-                            </label>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="actions">
-                <button type="submit" class="btn btn-primary">Save Notification Preferences</button>
-            </div>
-        </form>
-    </div>
-</div>
-
-{% endblock %}
