--- server/portal/apps/accounts/templates/portal/apps/accounts/manage_onboarding.html
@@ -1,9 +1,0 @@
-{% extends "portal/apps/accounts/index.html" %}
-
-{% block panel_content %}
-<div class="card">
-    <div class="card-body">
-        <onboarding-setup-view></onboarding-setup-view>
-    </div>
-</div>
-{% endblock %}
