--- server/portal/apps/accounts/templates/portal/apps/accounts/profile.html
@@ -1,54 +1,0 @@
-{% extends "portal/apps/accounts/index.html" %}
-{% block title %}Account Profile{% endblock %}
-{% block panel_content %}
-<div class="card">
-    <div class="card-body">
-        <h2>Account Profile</h2>
-        <hr>
-        <div class="row">
-            <div class="col-md-6">
-                <h3>Personal information</h3>
-                <dl>
-                    <dt>Full name</dt>
-                    <dd>{{profile.firstName}} {{profile.lastName}}</dd>
-
-                    <dt>Email address</dt>
-                    <dd>{{profile.email}}</dd>
-
-                    <dt>Phone number</dt>
-                    <dd>{{profile.phone}}</dd>
-
-                    <dt>Institution</dt>
-                    <dd>{{profile.institution}}{% if profile.department %},<br>
-                        {{profile.department}}{% endif %}
-                    </dd>
-
-                    <dt>Title</dt>
-                    <dd>{{profile.title}}</dd>
-
-                    <dt>Country of residence</dt>
-                    <dd>{{profile.country}}</dd>
-
-                    <dt>Country of citizenship</dt>
-                    <dd>{{profile.citizenship}}</dd>
-                </dl>
-            </div>
-            <!--
-            <div class="col-md-6">
-                <h3>Demographics</h3>
-                <dl>
-                    <dt>Ethnicity</dt>
-                    <dd>{{demographics.ethnicity|default:'<span class="text-warning">Please provide</span>'}}</dd>
-
-                    <dt>Gender</dt>
-                    <dd>{{demographics.gender|default:'<span class="text-warning">Please provide</span>'}}</dd>
-                </dl>
-            </div>
-            -->
-        </div>
-        <hr>
-        <a href="{% url 'portal_accounts:profile_edit' %}" class="btn btn-default"><i class="fas fa-edit"></i></i> Edit profile
-        </a>
-    </div>
-</div>
-{% endblock %}
