--- server/portal/apps/accounts/templates/portal/apps/accounts/register.html
@@ -1,46 +1,0 @@
-{% extends "base.html" %}
-{% load bootstrap4 sekizai_tags staticfiles %}
-{% block title %} Registration {% endblock %}
-{% block content %}
-  <div class="container">
-    <h1 class="headline headline-research">Register an Account</h1>
-
-    <div class="row">
-      <div class="col-md-5 col-md-push-7">
-        <div class="callout callout-default">
-          <p class="text-center">
-            <img style="max-width:300px" alt="TACC" src="{% static 'images/tacc-logo-outline.svg'%}">
-          </p>
-          <h3>
-            Already have an Account? <a href="{% url 'portal_auth:agave_oauth' %}">Log in.</a>
-          </h3>
-          <p>
-            If you already have a TACC account, simply
-            <a href="{% url 'portal_auth:agave_oauth' %}">log in</a> with your TACC username and password to access Portal.
-          </p>
-          <p>
-            You may have a TACC account if you have used XSEDE or TeraGrid in the past and had an allocation on a TACC resource.
-          </p>
-        </div>
-      </div>
-
-      <div class="col-md-7 col-md-pull-5">
-        <p class="lead">
-          Portal uses the TACC Identity Service. To use Portal you need to
-          register for a TACC Account. With a TACC Account you will be able to access
-          Portal as well as additional services provided by TACC. See the
-          <a href="https://www.tacc.utexas.edu" target="_blank">TACC Website</a> for more
-          information.
-        </p>
-        <form method="post" action="">
-          {% csrf_token %}
-          {% bootstrap_form account_form %}
-          <button class="btn btn-primary" type="submit">Register Account</button>
-          <a class="btn btn-link" href="/">Cancel</a>
-          <a class="btn btn-link" href="{% url 'login' %}">Already have an account? Log in.</a>
-        </form>
-      </div>
-    </div>
-  </div>
-
-{% endblock %}
