--- server/portal/apps/accounts/templates/portal/apps/accounts/request_access.html
@@ -1,26 +1,0 @@
-{% extends "base.html" %}
-{% load bootstrap4 sekizai_tags staticfiles %}
-{% block title %} Registration {% endblock %}
-{% block content %}
-  <div class="container">
-    <h2>Request Access to the Portal Platform</h2>
-    <p>At this time, access is restricted to Portal project members. Your username and password will be the same as your TACC username and password.</p>
-    <h4>Create a TACC account</h4>
-    <p>To create a TACC username and password, please visit this site: <a href="https://portal.tacc.utexas.edu/account-request" target="_blank">https://portal.tacc.utexas.edu/account-request</a></p>
-
-    <p>You must agree to the TACC Acceptable Use Policy, and then fill out the Account Request form.</p>
-    <p>After submitting the Account Request form, you should receive a confirmation email. Please go ahead and confirm.</p>
-    <p>Once your email is confirmed, your account request will then be reviewed. Account approval should occur within 1 business day. You should receive an email when your account is approved.</p>
-
-    <p>If you require further assistance, please submit a support ticket <a href="https://consult.tacc.utexas.edu/Ticket/Create.html?Queue={{rt_qn}}" target="_blank">here</a>.</p>
-
-    <form method="post" action="">
-        {% csrf_token %}
-        {% bootstrap_form access_form %}
-        <button class="btn btn-primary" type="submit">Request Access</button>
-        <a class="btn btn-link" href="/">Cancel</a>
-        <a class="btn btn-link pull-right" href="/accounts/login">Already have an account? Log in.</a>
-      </form>
-
-  </div>
-{% endblock %}
