+++ server/portal/apps/auth/views.py
@@ -131,10 +131,6 @@ def agave_oauth_callback(request):
             error = request.GET['error']
             logger.warning('Authorization failed: %s', error)
 
-        messages.error(request,
-                       'Authentication failed. Did you forget your password? '
-                       '<a href="%s">Click here</a> to reset your password.' %
-                       reverse('portal_accounts:password_reset'))
         return HttpResponseRedirect(reverse('portal_accounts:logout'))
 
     if 'next' in request.session:
